changing controllers location
